タブレット用のヘッダーコンポーネントを整理 #15
