--- src/components/TabletHeader/TabletSubMenuItem.jsx
@@ -1,15 +1,0 @@
-import { Link, ListItemIcon } from "@mui/material";
-import React from "react";
-
-const TabletSubMenuItem = (props) => {
-  return (
-    <>
-      <ListItemIcon>{props.subIcon}</ListItemIcon>
-      <Link href={props.route} sx={{ color: "black", textDecoration: "none" }}>
-        {props.subContent}
-      </Link>
-    </>
-  );
-};
-
-export default TabletSubMenuItem;
