+++ src/components/atoms/FormButton/FormButton.jsx
@@ -1,5 +1,4 @@
 import { Button, CircularProgress, Grid } from "@mui/material";
-import React from "react";
 
 const FormButton = ({ loading }) => {
   return (
