+++ src/components/atoms/ImageComponent/index.jsx
